Started on the ProjectManifest side of things.
A ProjectManifest records a dependency on a specific version of a project
so that we can extract and place that dependency in the parent project's
library directory.

Right now we look for manifests for all of the configured Project's. We
look in their Library directory (relative to the Build directory) for all
the projref files and just report the ones we find.