Removing dependencies on configuration.
I'm doing this by making the specific domain's richer, to include the
information they would usually pull from the configuration because it's
clear that information belongs as properties of the domain. For example
the ArchivePath of a Version should NOT be built the way it was, it's a
property of the Version and is thus populated by the repository.