Breaking archiving.
Right now we had a SourceLocation acting as the BuildDirectory in a
project and that's just kind of annoying because of how Location's work.

I want to navigate away from the Location's as static collections of
files that are turned into FileSet's for the various operations.

So this change will break the Project.MakeArchive for now until I find
a better way.