Got progress working with the Mirrored strategy.
Right now there's an event for these in DistributedDomainEvents and I'm
thinking that maybe that belongs in a more general place (or we make
something more specific as far as events go)

Right now it's number of files based, so for a few big files we'll look
terrible. That's something we can work on later.