Had to change when the repository was saved during publish/commit.
It was saved after the hook ran and so the new Manifest.xml was never
placed into the repository. This is just more reason to rethink how we
actually perform these mutable operations in the domain. We need to
group the side effects better and the whole thing should probably
really happen inside of the save.