Adding multiple types of ProjectReferences to express broken references.
When everything is fine and we can find all we need in the Repository
we create a HealthyProjectReference, otherwise we get a
BrokenProjectReference that has limited functionality and will only have
behavior for reporting the bad state so the user can take action. This
way we can continue to function w/o throwing so the user *can* fix the
problem with the reference.