ReferenceCandidate work.
As we move to multiple repositories we'll need to be able to distinquish
between projects with the same name in two different repositories. To
do this we're going to gather ReferenceCandidates and use the users
arguments and the configuration to build sensible defaults for the
project references they are referring to. For example, when adding a
reference if they don't specify a repository we can default the repository
name on the ReferenceCandidate to the only configured repository or null
if there's more than one.

We can basically work out how equality works a little easier this way.