Refactoring to make things easier to read.
Borrowed a technique from Aaron where you create a Mapping class with
extension methods to simplify going from one domain to another. I didn't
want to introduce this logic into the Core or Configuration domains
directly and the possibility of NULLs was making this logic ugly. This
is much pretty as we can hide the NULL test away in the extension method.