Working on moving more stuff into the application service layer.
This is work on manipulating the current project (adding dependencies) and
manipulating repositories (adding new versions) I also renamed a few things
and did some tweaking.

I adopted a *Response for the objects in the schema that comes back from
the ASL. We'll see how much we like this naming. It makes sense
because we're basically asking the application to do something.