Specificiations on adding new dependencies to projects.
The real problem with these is they don't really test the unzipping process
and that's probably more likely to go wrong.

Fixed some bugs in how dependencies/references are handled in the
collections that could cause us to get confused later if we ever
stopped doing just one-off changes in commands. (collections would get
duplicate entries and such) We now replace any existing reference for
that project with the new one that is being added.

Also made it so the repository name given when adding repositories is
used to disambiguate.