Introducing concept of a ProjectFromRepository.
This is just a reference of to an ArchivedProject and it's Repository. I
am just not happy with these names and I think I am waiting for a break-
through on how to capture these concepts in a useful manner. I'm thinking
of narrow down what's used on them via interfaces. This refactor cleans up
a lot though and is necessary to some future things.